Moves calendar styles and date utils to shared libs
Centralizes calendar CSS under shared directory for better maintainability and updates code to use common date formatting and parsing utilities.
Reduces duplication, clarifies style usage, and improves code organization.
