+++ src/app/request/create/travel/calendar-styles.css
@@ -1,145 +1,1 @@
-/* カレンダーのカスタムスタイル */
-.custom-calendar {
-  font-family: inherit;
-  width: 100%;
-  max-width: 400px;
-}
-
-.custom-calendar .rmdp-calendar {
-  border-radius: 12px;
-  box-shadow: 0 4px 6px -1px rgb(0 0 0 / 0.1);
-  width: 100%;
-  min-width: 320px;
-}
-
-.custom-calendar .rmdp-header {
-  border-radius: 12px 12px 0 0;
-  padding: 12px;
-}
-
-.custom-calendar .rmdp-week-day {
-  height: 32px;
-  line-height: 32px;
-  font-size: 14px;
-  font-weight: 600;
-}
-
-.custom-calendar .rmdp-day {
-  border-radius: 8px;
-  margin: 2px;
-  transition: all 0.2s ease;
-  height: 40px;
-  width: 40px;
-  font-size: 16px;
-  line-height: 40px;
-}
-
-.custom-calendar .rmdp-day:hover:not(.rmdp-disabled) {
-  background-color: #e0f2fe;
-  transform: scale(1.05);
-}
-
-.custom-calendar .rmdp-day.rmdp-selected {
-  background-color: #0ea5e9;
-  color: white;
-}
-
-.custom-calendar .rmdp-day.rmdp-disabled {
-  cursor: not-allowed !important;
-  opacity: 0.5;
-}
-
-/* 選択済み期間のスタイル（薄く表示） */
-.custom-calendar .rmdp-day.selected-range {
-  background-color: #fef2f2 !important;
-  color: #94a3b8 !important;
-  opacity: 0.8;
-  border: 1px solid #fecaca;
-}
-
-/* 完全一致の出発日（選択不可） */
-.custom-calendar .rmdp-day.selected-start-date-exact {
-  background-color: #ef4444 !important;
-  color: white !important;
-  cursor: not-allowed !important;
-  font-weight: bold;
-  border: 2px solid #dc2626;
-}
-
-/* 部分一致の出発日（他の期間で選択済み） */
-.custom-calendar .rmdp-day.selected-start-date-partial {
-  background-color: #f97316 !important;
-  color: white !important;
-  font-weight: bold;
-  border: 2px solid #ea580c;
-  position: relative;
-}
-
-.custom-calendar .rmdp-day.selected-start-date-partial::after {
-  content: '●';
-  position: absolute;
-  top: 2px;
-  right: 2px;
-  font-size: 8px;
-  color: white;
-}
-
-/* 過去の日付 */
-.custom-calendar .rmdp-day.past-date {
-  background-color: #f5f5f5 !important;
-  color: #ccc !important;
-  cursor: not-allowed !important;
-}
-
-/* 選択済み出発日のスタイル */
-.custom-calendar .selected-start-date {
-  background-color: #ef4444 !important;
-  color: white !important;
-  cursor: not-allowed !important;
-  position: relative;
-  font-weight: bold;
-}
-
-.custom-calendar .selected-start-date::after {
-  content: '×';
-  position: absolute;
-  top: 2px;
-  right: 4px;
-  font-size: 12px;
-  line-height: 1;
-}
-
-/* 選択済み期間のスタイル */
-.custom-calendar .selected-range {
-  background-color: #fee2e2 !important;
-  color: #666 !important;
-  border: 1px dashed #ef4444;
-}
-
-/* 過去の日付のスタイル */
-.custom-calendar .past-date {
-  background-color: #f5f5f5 !important;
-  color: #ccc !important;
-  cursor: not-allowed !important;
-}
-
-/* レスポンシブ対応 */
-@media (max-width: 640px) {
-  .custom-calendar .rmdp-calendar {
-    width: 100%;
-    min-width: 300px;
-  }
-
-  .custom-calendar .rmdp-day {
-    font-size: 14px;
-    height: 36px;
-    width: 36px;
-    line-height: 36px;
-  }
-
-  .custom-calendar .rmdp-week-day {
-    height: 28px;
-    line-height: 28px;
-    font-size: 12px;
-  }
-}
+/* このファイルは非推奨です。スタイルは /src/lib/react-multi-date-picker/calendar.css に移動されました */
