+++ src/lib/react-multi-date-picker/Calendar.tsx
@@ -5,6 +5,8 @@ import {
   CalendarProps as ReactMultiDatePickerCalendarProps,
 } from 'react-multi-date-picker';
 
+import './calendar.css';
+
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
 interface CalendarProps extends ReactMultiDatePickerCalendarProps {}
